requirements.txt

fastapi
uvicorn
pymupdf
pillow
pytesseract
requests
